Time: 9 ms (8.07%), Space: 8.7 MB (40.46%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -2,10 +2,10 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         int num=x;
-        int x1=0;
+        long long x1=0;
         while(x>0){
             int rem=x%10;
-            x1=x1*10+rem;
+            x1=(long long)x1*10+rem;
             x/=10;
         }
         return (x1==num);
